enable running via -m and remove old test
